+++ quantify_scheduler/backends/qblox/operations/pulse_library.py
@@ -20,22 +20,32 @@ class LatchReset(Operation):
         self,
         t0: float = 0,
         duration: float = 4e-9,
-        port: str | None = None,
-        clock: str | None = None,
+        portclocks: list[tuple[str, str]] | None = None,
     ) -> None:
         super().__init__(name=self.__class__.__name__)
+        if portclocks is None:
+            portclocks = []
         self.data["pulse_info"] = [
             {
                 "name": self.__class__.__name__,
                 "wf_func": None,
                 "t0": t0,
-                "clock": clock,
-                "port": port,
+                "port": portclock[0],
+                "clock": portclock[1],
                 "duration": duration,
             }
+            for portclock in portclocks
         ]
         self._update()
 
     def __str__(self) -> str:
         pulse_info = self.data["pulse_info"][0]
-        return self._get_signature(pulse_info)
+        return (
+            f"{self.__class__.__name__}("
+            f"name='{self.name}', "
+            f"t0='{pulse_info['t0']}', "
+            f"port='{pulse_info['port']}'"
+            f"clock='{pulse_info['clock']}'"
+            f"duration='{pulse_info['duration']}'"
+            f")"
+        )